use uint8_t for inv_palette[]

+++ src/enc/vp8l.c
@@ -825,7 +825,7 @@ static void ApplyPalette(uint32_t* src, uint32_t* dst,
   }
 
   if (use_LUT) {
-    int inv_palette[MAX_PALETTE_SIZE] = { 0 };
+    uint8_t inv_palette[MAX_PALETTE_SIZE] = { 0 };
     for (i = 0; i < palette_size; ++i) {
       const int color = (palette[i] >> 8) & 0xff;
       inv_palette[color] = i;
